requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
werkzeug>=3.0.0
openai>=2.2.0
moviepy>=1.0.3
pillow>=11.3.0
requests>=2.31.0
python-dotenv>=1.0.0
gunicorn>=21.2.0
python-jose[cryptography]>=3.3.0

